agents/financial/requirements.txt
# Ledger Financial Agent Dependencies
requests>=2.28.0
python-dateutil>=2.8.0
